~~ Use the new commonization algorithm
Revert the now-unnecessary changes.

+++ ktor-client/ktor-client-darwin-legacy/darwin/src/io/ktor/client/engine/darwin/internal/legacy/DarwinLegacyUtils.kt
@@ -45,7 +45,7 @@ internal fun ByteArray.toNSData(): NSData = NSMutableData().apply {
 
 @OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
 internal fun NSData.toByteArray(): ByteArray {
-    val result = ByteArray(length.toULong().toInt())
+    val result = ByteArray(length.toInt())
     if (result.isEmpty()) return result
 
     result.usePinned {
@@ -61,7 +61,7 @@ internal fun NSData.toByteArray(): ByteArray {
  */
 @OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
 internal suspend fun ByteWriteChannel.writeFully(data: NSData) {
-    val length = data.length.toULong().toLong()
+    val length = data.length.toLong()
     if (length == 0L) return
 
     val bytes = data.bytes ?: return
