+++ ktor-io/darwin/src/CharsetDarwin.kt
@@ -110,7 +110,7 @@ private fun ByteArray.toNSData(): NSData = NSMutableData().apply {
 
 @OptIn(UnsafeNumber::class, ExperimentalForeignApi::class)
 private fun NSData.toByteArray(): ByteArray {
-    val result = ByteArray(length.toULong().toInt())
+    val result = ByteArray(length.toInt())
     if (result.isEmpty()) return result
 
     result.usePinned {
